v1.0
Extension works! Match and team info is now worked out in the background.js file - so much faster loading times. New scores are sent to the extension every 10 seconds (assuming it is still loaded).
List of live matches is refreshed every 60 seconds in the background - this check for new live matches and will add them to the array of matches should one be found.

Content.js is a required file, even if empty.

Some small bugs that need to be ironed out.